Revert custom update
